update play 2.5.0
