+++ play/src/main/scala/com/github/levkhomich/akka/tracing/play/PlayActorTracing.scala
@@ -17,10 +17,9 @@
 package com.github.levkhomich.akka.tracing.play
 
 import akka.actor.Actor
-import play.api.Routes
 import play.api.mvc.RequestHeader
-
-import com.github.levkhomich.akka.tracing.{ TracingExtensionImpl, BaseTracingSupport, ActorTracing }
+import com.github.levkhomich.akka.tracing.{ ActorTracing, BaseTracingSupport, TracingExtensionImpl }
+import play.routing.Router
 
 trait PlayActorTracing extends ActorTracing { self: Actor =>
 
@@ -35,7 +34,7 @@ class PlayRequestTracingSupport(val headers: RequestHeader) extends AnyVal with 
     headers.id
 
   override protected[tracing] def spanName: String = {
-    val route = headers.tags.getOrElse(Routes.ROUTE_PATTERN, headers.path)
+    val route = headers.tags.getOrElse(Router.Tags.ROUTE_PATTERN, headers.path)
     headers.method + " " + route
   }
